:sparkles: Feat: Implementa modal de docs na view comercial
# Descrição

Um botão similar ao que abre o sheet foi implementado na tabela de `Documentação a aprovar`, abrindo um modal com os documentos cadastrados, como na view de broker

---

### Outras alterações

Os componentes de tabelas de `Documentação a aprovar` e `Pendente de enviar para Due` sofreram mudanças nos tamanhos das suas células e das suas colunas, recebendo também alguns placeholders em lugares onde os dados chegam null/undefined
