+++ src/components/Modals/BrokersDocs/index.tsx
@@ -1,4 +1,6 @@
 "use client";
+import React from "react";
+import ReactDOM from "react-dom";
 import { BrokersContext } from '@/context/BrokersContext';
 import { useContext, useEffect, useState } from 'react';
 import { Fade } from 'react-awesome-reveal';
@@ -36,10 +38,10 @@ const DocForm = () => {
         setCredorIdentificationType(credorIdent.length === 11 ? "CPF" : credorIdent.length === 14 ? "CNPJ" : null);
     }, [docModalInfo]);
 
-    return (
+    return ReactDOM.createPortal (
       <div
         role="dialog"
-        className="fixed left-0 top-0 z-999 flex h-screen w-screen animate-fade items-center justify-center bg-black-2/50 bg-opacity-10 bg-clip-padding backdrop-blur-sm backdrop-filter transition-all duration-300 ease-in-out"
+        className="fixed left-0 top-0 z-999 flex h-screen w-screen animate-fade items-center justify-center bg-[#00000040] bg-opacity-10 bg-clip-padding backdrop-blur-sm backdrop-filter transition-all duration-300 ease-in-out"
       >
         <div className="relative h-fit rounded-lg border border-stroke bg-snow dark:border-strokedark dark:bg-boxdark-2 2xsm:w-11/12 2xsm:py-8 2xsm:px-5 md:w-4/5 md:p-10 lg:w-10/12 3xl:w-10/12">
           <button className="group absolute right-2 top-2 flex h-8 w-8 cursor-pointer items-center justify-center rounded-full transition-colors duration-300 hover:bg-slate-700">
@@ -87,7 +89,8 @@ const DocForm = () => {
             </>
           )}
         </div>
-      </div>
+      </div>,
+      document.body
     );
 }
